+++ kuiper/include/sampler/sampler.h
@@ -10,7 +10,7 @@ public:
 
     // argmax 贪心采样: 模型输出的概率分布 logits，其长度为 size，输出概率最高的那个 token 的索引
     // 输入 logits = [-0.2, 2.3, 0.5, 1.8, -1.0]，输出为索引 1
-    virtual size_t sample(const float* logits, size_t size, void* stream = nullptr) const = 0;
+    virtual int32_t sample(const float* logits, int32_t size, void* stream = nullptr) const = 0;
 
 protected:
     base::DeviceType device_type_ = base::DeviceType::DeviceUnknown;
