+++ kuiper/source/op/kernel/cuda/argmax_kernel.cu
@@ -1,7 +1,7 @@
 #include "argmax_kernel.cuh"
 
 namespace kernel {
-size_t argmax_kernel_cu(const float* input_ptr, size_t size, void* stream) {
+int32_t argmax_kernel_cu(const float* input_ptr, int32_t size, void* stream) {
     return 0;
 }
 }  // namespace kernel
